+++ Project/WinAPI/slimeBlue.h
@@ -4,7 +4,7 @@
 class SlimeBlue : public Enemy
 {
 private:
-	float turnCount;
+	double turnCount;
 	float attackDelay;
 	float atkRange;
 	float fieldOfVision;
